No fragment bundles
An assumption made in advance states that fragment bundles are assumed. Tests show that this instrument is not required for bolt or http driver transport.

Maven properties

Cleanup maven properties in feature modules, handle props in parent module. Introduce property neo4j-ogm.version to reflect which version the neo4j-ogm as a git submodule has.