Add dedicated Neo4j OGM OSGi module
The purpose of a dedicated module is to provide OSGi-based components, or factories for the SessionFactory Neo4j-OGM provides, a real OSGi-based application can consume.

Most important is the class-loading aspect a real application should not have to tamper with.

Both classes OGM-Session-Component and OGM-SessionFactory switch the class-loader used to generate Neo4j-OGM-SessionFactory instance, see the class Class-Loader-Switcher for more detail.

The user chooses among OGM-Session-Component or OGM-SessionFactory in its application. Examples of how to use those got illustrated in detail in the documentation, respectively, in the demo application provided.

The Session-Factories configuration is controllable in an OSGI-compatible way, more also in the documentation.

The class-loading strategy applied by Neo4j-OGM is to use the current threads context class-loader most of the time.

Thread.currentThread().getContextClassLoader()

The class-loader gets used then as a parameter for Javas API Class.forName.

This strategy is mostly not compatible with OSGi, where only the simplest version of Class.forName works 100%, of course only when clear contracts among API and implementation are guaranteed. Studying the Neo4j-OGM codebase illustrates a large room for improvements in this area.

One problem is that class loading is not optioned, meaning that environments adapted to a different class-loading strategy got not supported out of the box, and OSGi is no exception.

Testing shows that Equinox, as an OSGi runtime has somehow worked around the problem, the class loader of the current thread works out of the box with this strategy.

In an OSGi-based environment, this is just the tip of the iceberg. Handling fragment-bundles to fuse things using a single class-loader certainly complicates things here and there.

This is just a reminder when looking at the changes to determine the purpose; for more information, look at the documentation.