Replace bundle-plugin with bnd-plugin
Replacing the plugin required a modification of how bnd-files from a resource directory are bound served to the individual modules.

With the maven bundle plugin, it was possible to use a maven property session provided "executionRootDirectory" to reference a resource folder where bnd-files live relative on the base folder.

To recreate this behavior, the directory-maven plugin gets used. The directory plugin requires a phase earlier or at the same time as the bnd-maven plugin gets executed, which is process-classes. The bnd-maven plugin uses a bndfile with the title "delegate.bnd" as a wrapper to load/include the respective bnd-file per processed module.

Change Fragment-Host declaration
org.neo4j.ogm-core => neo4j-ogm-core only a modules artifactId gets used as the symbolic name of the bundle.

Each bnd descriptor file requires a more specific declaration which packages get exported, versions for shipped packages get self-generated.