Remove BND Feature to autogen JPMS declarations
There is a conflict with mavens surefire-plugin. It is not possible to run the surefire and bnds option "-jpms-module-info:" at the same time, at least not without the option "forkCount 0" set. For now, the jpms build option gets disabled. In the future, we have to figure out what is the cause of this problem.

Add missing fragment host declarations.